+++ dockin-opsctl/internal/ssh/terminal_windows.go
@@ -1,0 +1,21 @@
+/*
+ * Copyright (C) @2021 Webank Group Holding Limited
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+// +build windows
+package ssh
+
+// on windows just for mock, as no api to catch the terminal size change
+// maybe later found it
+func UpdateTerminalSize(winSize chan WindowSize, exitChan chan byte) {
+}
